bot/requirements.txt
aiogram==3.23.0
aiogram_dialog==2.4.0
aiohttp==3.13.2
python-dotenv==1.2.1
requests==2.32.5
redis==7.1.0